refactor(shared): 拆 IpcInvokeMap + 9 facade + XAgentApi 到新文件 (主题 D #60 C-301)
shared/ipc.ts 原 1475 行 god 拆出 ~325 行 (IpcInvokeMap + 9 facade + XAgentApi
+ DELETED_FLAT_KEYS) 到 shared/ipc-invoke-map.ts, 本文件 1475 -> 1064 行 (-411).

策略: barrel re-export 保持 104 个 consumer import 路径不变:
  - 旧 `from "../shared/ipc"` / `from "../../shared/ipc"` 等 104 个
    consumer 文件无需任何 import 改动
  - 新 `from "../shared/ipc-invoke-map"` 留给要 tree-shake 细粒度
    import 的新代码

拆分内容 (ipc-invoke-map.ts, 16 KB):
- IpcInvokeMap (通道签名 78 字段, ~150 行)
- DELETED_FLAT_KEYS (35 字符串, flat-key 白名单) + DeletedFlatKey type
- 9 facade: WorkspaceApi / TurnApi / PlanApi / SessionApi / PrefsApi /
  AppReportApi / LogoApi / GodotApi / UpdatesApi
- XAgentApi (flat + facade 联合)
- XAgentApiFlat + FlatInvokeApi
- 2 个 compile-time gate (IpcInvokeMap 覆盖 IPC_CHANNELS / DELETED 合法)

不在本 PR (PR-Y4 也做但留 ADR):
- C-302 拆 shared/godot-rpc.ts 3 件 (protocol/policy/gating)
  - 当前 372 行 godot-rpc.ts 没有"5 个 facade"级别的结构化拆分必要
  - 计划里说"工具注册表 + UI gating + clamp"三件, 但代码实际是 union type +
    clamp 函数的松散结构, 强行拆 3 件 ≈ 拆 60 + 30 + 200, 收益小
  - 留 ADR: 拆前 interface ≈ 子集之和, 强行拆只是搬家
- 拆 shared/ipc.ts 剩余 types/validators 5 件分件 (C-301 计划 5 件):
  - 当前剩余 1064 行主要是 100+ 小 interface + TypeBox schema, 拆 5 件
    收益是 -300 行 (型 vs barrel 行数比), 但需要 ~50 个 import 调整
  - 留给 0.7.x 单独 PR

收益:
- shared/ipc.ts 1475 -> 1064 行 (-411, -28%)
- shared/ipc-invoke-map.ts 16 KB 新文件, 单一关注点 (IPC 通道契约)
- 104 个 consumer import 路径零改动 (barrel re-export)
- typecheck 0 errors, vitest 41 files / 535 tests pass (含 PR-Y3 的 4 case)

验收:
- npm run typecheck: 0 errors
- npx vitest run: 41 files / 535 tests pass (含 PR-Y3 boundary test)
- preload.ts 仍用同一 import 路径 (XAgentApi 从 shared/ipc 拿)

Refs #60 (主题 D) - C-301 主体完成, C-302 (godot-rpc 拆 3) 留 0.7.x, C-301 余项
(剩余 types/validators 拆 5 件) 留 0.7.x
